Create dedicated individual resource types
With the current `value` and `type` style generic infoblox_record
resource it is very difficult to cleanly add other resources to the
provider. There is enough variety to DNS record types that having
required `domain` and `name` attributes does not always make sense.
For example an MX, SRV, and CNAME record don't have much in common
with each other. This commit is extracts the currently supported
records in the 'infoblox_record' resource type to their own resources
and adds the PTR type.